Example target data

Example target data consisting of incident influenza hospitalization admissions in the US from 2020-2023. This data can be useful as a modeling input and for plots of time series values.

Usage

example_target_data

Format

example_target_data

A data frame with 10,255 rows and 3 columns:

time_idx

the Saturday ending the week during which hospital admission were recorded, in format yyyy-mm-dd

location

FIPS code identifying a location

value

the number of influenza hospital admissions in the specified location over the course of the week ending on the specified date, time_idx

...

Source

https://github.com/Infectious-Disease-Modeling-Hubs/example-complex-forecast-hub/